Added http2_test.go
